added extension & docercompose

+++ exercise.wwwapi/Repository/Repository.cs
@@ -13,41 +13,57 @@ namespace exercise.wwwapi.Repository
             _db = db;
         }
         
-        /*
         public async Task<IEnumerable<Course>> GetCourses()
         {
             return await _db.Courses.ToListAsync();
         }
-        */
-
+        
         public async Task<IEnumerable<Student>> GetStudents()
         {
-            return await _db.Students.ToListAsync();
+            return await _db.Students.Include(s => s.Course).ToListAsync();
         }
 
-        public async Task<Student> CreateStudent(CreateNewStudentPayload createData)
+        public async Task<Student> CreateStudent(StudentPayload createData)
         {
+            if(
+                createData.FirstName == null ||
+                createData.LastName == null ||
+                createData.DateOfBirth == null)
+            {
+                return null;
+            }
+
             Student student = new Student();
             student.FirstName = createData.FirstName;
             student.LastName = createData.LastName;
             student.DateOfBirth = createData.DateOfBirth;
-            student.CourseTitle = createData.CourseTitle;
-            student.CourseStartDate = createData.CourseStartDate;
             student.AverageGrade = createData.AverageGrade;
+            student.CourseId = createData.courseId;
             _db.Students.Add(student);
             _db.SaveChanges();
             return student;
         }
 
-        public async Task<Student> UpdateStudent(int id, UpdateStudentPayload updateData)
+        public async Task<Student> UpdateStudent(int id, StudentPayload updateData)
         {
             var student = await _db.Students.FindAsync(id);
+            if(student == null)
+            {
+                return null;
+            }
+            if (
+                updateData.FirstName == null ||
+                updateData.LastName == null ||
+                updateData.DateOfBirth == null)
+            {
+                return null;
+            }
+
             student.FirstName = updateData.FirstName;
             student.LastName = updateData.LastName;
             student.DateOfBirth = updateData.DateOfBirth;
-            student.CourseTitle = updateData.CourseTitle;
-            student.CourseStartDate = updateData.CourseStartDate;
             student.AverageGrade = updateData.AverageGrade;
+            student.CourseId = updateData.courseId;
             _db.SaveChanges();
             return student;
         }
@@ -55,6 +71,10 @@ namespace exercise.wwwapi.Repository
         public async Task<Student> DeleteStudent(int id)
         {
             var student = await _db.Students.FindAsync(id);
+            if(student == null)
+            {
+                return null;
+            }
             _db.Students.Remove(student);
             _db.SaveChanges();
             return student;
